requirements.txt
python-dotenv
numpy
pandas
matplotlib
requests
beautifulsoup4
lxml
yfinance
pyarrow
